Expose `cpl_normalise_missing()` in `type-complex.h`
